+++ app/controllers/keys_controller.rb
@@ -3,7 +3,7 @@ class KeysController < ApplicationController
   before_action :authorize
   before_action :find_key, only: [ :show, :edit, :update, :destroy, :copy ]
   before_action :find_keys, only: [ :context_menu ]
-  accept_api_auth :index, :show
+  accept_api_auth :index, :show, :create, :update, :destroy
 
   helper :sort
   include SortHelper
@@ -33,7 +33,7 @@ class KeysController < ApplicationController
     if @query
       if @query.match(/#/)
         tag_string = (@query.match(/(#)([^,]+)/))[2]
-        tag = Vault::Tag.find_by_name(tag_string)
+        tag = @project.vault_tags.find_by(name: tag_string)
         @keys = tag.nil? ? nil : tag.keys.where(project: @project)
       else
         if params[:search_fild] == 'name'
@@ -41,7 +41,7 @@ class KeysController < ApplicationController
         elsif params[:search_fild] == 'url'
           @keys = @project.keys.where(url: @query)
         elsif params[:search_fild] == 'tag'
-          tag = Vault::Tag.find_by_name(@query)
+          tag = @project.vault_tags.find_by(name: @query)
           @keys = tag.nil? ? nil : tag.keys.where(project: @project)
         end
       end
@@ -67,7 +67,7 @@ class KeysController < ApplicationController
     respond_to do |format|
       format.html
       format.pdf
-      format.json { render json: @keys }
+      format.api { render json: @keys.map { |k| key_json(k) } }
     end
   end
 
@@ -88,15 +88,17 @@ class KeysController < ApplicationController
 
     @key.project = @project
 
-    @key.tags = Vault::Tag.create_from_string(tags_param)
+    @key.tags = Vault::Tag.create_from_string(tags_param, @project)
 
     self.update_wishlist
 
     respond_to do |format|
       if @key.save
         format.html { redirect_to project_keys_path(@project), notice: t('notice.key.create.success') }
+        format.api  { render json: key_json(@key), status: :created, location: project_key_url(@project, @key) }
       else
         format.html { render action: 'new' }
+        format.api  { render json: { errors: @key.errors.full_messages }, status: :unprocessable_entity }
       end
     end
   end
@@ -108,10 +110,12 @@ class KeysController < ApplicationController
       self.update_wishlist
 
       if @key.update(key_params)
-        @key.tags = Vault::Tag.create_from_string(tags_param)
+        @key.tags = Vault::Tag.create_from_string(tags_param, @project)
         format.html { redirect_to project_keys_path(@project), notice: t('notice.key.update.success') }
+        format.api  { render json: key_json(@key.reload.decrypt!), status: :ok }
       else
         format.html { render action: 'edit'}
+        format.api  { render json: { errors: @key.errors.full_messages }, status: :unprocessable_entity }
       end
     end
   end
@@ -140,20 +144,29 @@ class KeysController < ApplicationController
 
   def show
     if !@key.whitelisted?(User,@project)
-      render_error t("error.key.not_whitelisted")
+      respond_to do |format|
+        format.html { render_error t("error.key.not_whitelisted") }
+        format.api  { render json: { errors: [t("error.key.not_whitelisted")] }, status: :forbidden }
+      end
       return
     else
       @key.decrypt!
       respond_to do |format|
         format.html { render action: 'show'}
+        format.api  { render json: key_json(@key) }
       end
     end
   end
 
   def destroy
-    Vault::Key.find(params[:id]).destroy
-    redirect_to project_keys_path(@project)
-    flash[:notice] = t('notice.key.delete.success')
+    @key.destroy
+    respond_to do |format|
+      format.html do
+        redirect_to project_keys_path(@project)
+        flash[:notice] = t('notice.key.delete.success')
+      end
+      format.api { head :no_content }
+    end
   end
 
   def context_menu
@@ -173,17 +186,37 @@ class KeysController < ApplicationController
 
   def find_keys
     @keys=Vault::Key.find(params[:ids])
-    unless @keys.all? { |k| k.project_id == @project.id } 
+    unless @keys.all? { |k| k.project_id == @project.id }
       redirect_to project_keys_path(@project), notice: t('alert.key.not_found')
     end
   end
 
   def key_params
-    params.require(:vault_key).permit(:type, :name, :body, :login, :file, :url, :comment, :tags).except(:tags)
+    params.require(:vault_key).permit(:type, :name, :body, :login, :file, :url, :comment).except(:tags)
   end
 
+  # Tags may arrive as a comma-separated string (web form) or as an array (API).
+  # Normalise to a string for Vault::Tag.create_from_string.
   def tags_param
-    params.require(:vault_key).permit(:tags)[:tags]
+    raw = params.dig(:vault_key, :tags)
+    raw.is_a?(Array) ? raw.join(', ') : raw
+  end
+
+  # Serialise a (decrypted) key for API responses, exposing tag names rather
+  # than join-table ids.
+  def key_json(key)
+    {
+      id: key.id,
+      project_id: key.project_id,
+      type: key.type,
+      name: key.name,
+      login: key.login,
+      body: key.body,
+      url: key.url,
+      comment: key.comment,
+      file: key.file,
+      tags: key.tags.map(&:name)
+    }
   end
 
   def index_params